Add the ability to register custom built-in resource packs. These packs are mainly used for adding additional mod support that is not dependent on java code.
These packs are disabled by default, this is to keep the functionality of allowing the end user of disabling the packs if they wanted to.

Waiting the following MinecraftForge issue to be resolved to allow these packs to be enabled by default while still allowing end user to disable them.
https://github.com/MinecraftForge/MinecraftForge/issues/8960